requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
codecov==2.1.10
constantly==15.1.0
coverage==5.3
cryptography==3.2.1
daphne==3.0.1
distlib==0.3.1
Django==3.1.3
django-admin-interface==0.14.1
django-apptemplates==1.5
django-ckeditor==6.0.0
django-colorfield==0.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-uuslug==1.2.0
filelock==3.0.12
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-slugify==4.0.1
pytz==2020.4
requests==2.25.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tox==3.20.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
virtualenv==20.1.0
zope.interface==5.2.0
